--- lib/client.ts
@@ -1,4 +1,0 @@
-import { PrismaClient } from "@prisma/client";
-
-let dbClient = new PrismaClient()
-export default dbClient;
