+++ pages/api/xAPI/statements/index.ts
@@ -3,7 +3,7 @@ import deepEqual from "deep-equal";
 import type { NextRequest, NextResponse } from "next/server";
 import { apiHandler } from "../../helpers/api/api-handler";
 import middleware, { Next } from "../../helpers/api/request-sanitizers";
-import dbClient from "../../../../lib/client";
+import dbClient from "../../../../lib/db";
 import { Prisma } from "@prisma/client";
 
 const helpers = middleware();
