+++ test.sh
@@ -1,5 +1,11 @@
 export IS_CI=${1}
 
-docker-compose down --remove-orphans
+function cleanup {
+	docker-compose down --remove-orphans
+}
+
+trap cleanup EXIT
+cleanup
+
 docker-compose build main
-docker-compose --env-file .env.test run main
+docker-compose run main
